==> notebooks/02_plotProportions.Rmd <==
---
title: "Plot the proportions"
author:
  - name: "Tom Smith"
    affiliation: "Cambridge Centre for Proteomics"
date: "`r format(Sys.time(), '%d %B, %Y')`"
abstract: | 
  Here we use the protein-level propertion object rendered in the previous notebook 
  and visualize the change in incorporation over time in the different samples.

output:
  pdf_document:
  html_notebook: default
geometry: margin=1in
fontfamily: mathpazo
fontsize: 11pt
---


```{r}
suppressMessages(library(dplyr))
suppressMessages(library(tidyr))
suppressMessages(library(ggplot2))
suppressMessages(library(MSnbase))
suppressMessages(library(camprotR))
suppressMessages(library(here))

# Standardise plotting aesthetics
source(here('R/plotting.R'))

```



```{r}
prot_proportions <- readRDS(here('results/prot_proportion.rds'))
pep_proportions_counts <- readRDS(here('results/pep_proportion_long.rds')) %>%
  filter(is.finite(ratio)) %>%
  group_by(Master.Protein.Accessions, replicate, condition) %>%
  tally()
```

Prepare proportions for plotting
```{r}
get_prot_proportions_for_plot <- function(prot_proportions){
  
  prot_proportions_for_plot <- prot_proportions %>%
    exprs() %>%
    merge(fData(prot_proportions), by='row.names') %>%
    as.data.frame() %>%
    gather(key='sample', value='proportion',
           -c(Row.names, Master.Protein.Descriptions)) %>%
    merge(pData(prot_proportions), by.x='sample', by.y='row.names') %>%
    mutate(proportion=as.numeric(proportion)) %>%
    filter(is.finite(proportion)) %>%
    mutate(type=factor(type, levels = c("3", "6", "24")))
  
  return(prot_proportions_for_plot)
}

prot_proportions_for_plot <- get_prot_proportions_for_plot(prot_proportions)

head(prot_proportions_for_plot)

```


Plot proportions
```{r}

p <- prot_proportions_for_plot %>%
  mutate(condition=recode(condition, 'untrt'='Control')) %>%
  ggplot(aes(proportion, colour=condition, group=interaction(condition, replicate))) +
  geom_density() +
  theme(axis.text=element_text(size=10)) +
  scale_colour_manual(values=c(get_cat_palette(2)), name='') +
  xlab('Proportion nascent protein') +
  facet_grid(type~.) +
  ylab('Density') +
  scale_x_continuous(limits=c(0,1), breaks=seq(0,1,0.2))
      
    
print(p)
ggsave(here('notebooks/plots/proportion_nascent_dist.png'))
ggsave(here('notebooks/plots/proportion_nascent_dist.pdf'))

```


```{r}
p <- prot_proportions_for_plot %>%
  mutate(condition=recode(condition, 'untrt'='Control')) %>%
  ggplot(aes(interaction(replicate, condition), proportion, colour=condition)) +
  geom_violin() +
  geom_boxplot(width=0.2, colour='black', outlier.shape = NA) +
  theme(axis.text=element_text(size=10)) +
  scale_colour_manual(values=c(get_cat_palette(3), 'black'), name='') +
  ylab('Proportion nascent protein') +
  facet_grid(.~type, scales='free', space='free') +
  xlab('') +
  theme(axis.text.x=element_text(angle=45, vjust=1, hjust=1))
      
    
print(p)
ggsave(here('notebooks/plots/proportion_nascent_dist_violin.png'))
ggsave(here('notebooks/plots/proportion_nascent_dist_violin.pdf'))
```
```{r, fig.height=6, fig.width=3}
p <- prot_proportions_for_plot %>%
  mutate(condition=recode(condition, 'untrt'='Control')) %>%
  arrange(type) %>%
  mutate(timepoint=sprintf('%s hr', type)) %>%
  mutate(timepoint=factor(timepoint, levels=unique(timepoint))) %>%
  ggplot(aes(proportion, colour=condition, group=interaction(condition, replicate))) +
  stat_density(geom="line",position="identity") +
  theme_camprot(base_size=15, border=FALSE, base_family='sans') +
  scale_colour_manual(values=get_cat_palette(2), name='') +
  xlab('Proportion nascent\nprotein') +
  facet_wrap(~timepoint, ncol=1) +
  ylab('Density') +
  scale_x_continuous(limits=c(0,1), breaks=seq(0,1,0.5)) +
  theme(strip.background=element_blank(),
        legend.background=element_blank(),
        legend.position=c(0.7,0.97))
  

print(p)
ggsave(here('figures/plots/proportion_nascent_dist.pdf'))
ggsave(here('figures/plots/proportion_nascent_dist.png'))

```

```{r}
compare_proportions_conditions <- prot_proportions_for_plot %>%
      select(-sample) %>%
      pivot_wider(values_from=proportion, names_from=condition)

saveRDS(compare_proportions_conditions, here('results/compare_proportions_conditions.rds'))

compare_proportions_type <- prot_proportions_for_plot %>%
      select(-sample) %>%
      pivot_wider(values_from=proportion, names_from=type)

saveRDS(compare_proportions_type, here('results/compare_proportions_type.rds'))
  



```




Plot correlation of proportions between DMSO & Tg. Median correlation here and overall lower nascent in Tg, especially for RNA-bound, as expected.
```{r, fig.height=7, fig.width=7}
condition_proportion_cor_per_rep <- compare_proportions_conditions %>%
  filter(is.finite(GC7), is.finite(untrt)) %>%
  group_by(type, replicate) %>%
  summarise(cor=cor(GC7, untrt, use='complete', method='spearman'))
  
compare_proportions_conditions %>%
  ggplot(aes(y=100*GC7, 100*untrt)) +
  xlim(0, 100) +
  ylim(0, 100) +
  geom_abline(slope=1, linetype=2, colour='grey') +
  geom_point(size=0.5, alpha=0.5) +
  geom_text(data=condition_proportion_cor_per_rep, inherit.aes=FALSE,
            aes(label=paste0('rho=', sprintf("%.2f", cor))), x=0, y=100, hjust=0, size=3) +
  facet_grid(replicate~type) +
  ylab('Proportion nascent in GC7') +
  xlab('Proportion nascent in untrt') +
  theme(text=element_text(size=15))
```   
OK, great, so there is a good correlation between the conditions, but at 3hr there seem to be more proteins off the diagonal (higher or lower translation with GC7 inhibition). I suspect this is just less accurate quantification. Later, we'll see if these are consistent differences.


Below, we plot a PCA for the proportions. Note, the major souce of variance is the time, since we will enevitably have more nascent protein at 24h than 3h or 6h. The effect of GC7 is 1-2 orders of magnitude less than the difference between the time points.

```{r, fig.height=5, fig.width=8}
pca <- prot_proportions %>%
  filterNA() %>%
  exprs() %>%
  t() %>%
  prcomp(scale=TRUE, centre=TRUE)

pov <- summary(pca)$importance[2,]

p <- pca$x %>%
  data.frame() %>%
  tibble::rownames_to_column('sample') %>%
  separate(sample, into=c('Time', 'Treatment', 'replicate')) %>%
  mutate(Time=factor(Time, levels=c(3,6,24))) %>%
  mutate(Treatment=recode(Treatment, 'untrt'='Control')) %>%
  ggplot(aes(PC1, PC2, fill=Time, shape=Treatment)) +
  geom_point(size=5) +
  scale_shape_manual(values=c(21,25)) +
  scale_fill_manual(values=get_cat_palette(3),
                    guide = guide_legend(override.aes = list(size = 6, shape = 21)),
                    name='Time (h)') +
  xlab(sprintf('PC1 (%s%%)', 100*round(pov[1], 2))) +
  ylab(sprintf('PC2 (%s%%)', 100*round(pov[2], 2))) +
  theme(aspect.ratio=1)

print(p)

ggsave(here('figures/plots/pca_proportions.pdf'))
ggsave(here('figures/plots/pca_proportions.png'))
  
```
